Refactor dmypy to move daemonize and Server to a separate file
